mover logica a appcurriculos
